PyTorch/requirements.txt
pytorch==2.1.2
torchmetrics==1.3.1
pandas==2.0.3
pathlib==1.0.1
matplotlib==3.7.2
scikit-learn==1.4.1.post1
mlxtend==0.23.1
